feat: switch to debian based build
When trying to upgrade arcgis to latest, the version of pyarrow that it requires was not building properly like before, and the version of pyarrow it requires doesn't have musl based wheels available. Don't want to speed anymore time trying to get it to build on alpine. I tried to replicate this workaround (https://github.com/apache/arrow/issues/39846#issuecomment-2632109524) in poetry but didn't work so I give up